ENH: Add code format check `GitHub` action workflow file
Add a code formatting check `GitHub` action workflow file.

As a consequence, remove the `flake8` code checking task from the `test,
package` workflow file.

Take advantage of the commit to exclude `GitHub` actions workflow files from
being checked by the `check-yaml` hook due to the fact that these files might
have some expressions to set variables (e.g. `${{ <expression> }}`) that are not
correctly interpreted by the hook.

Add the corresponding badge to the `README` file.

Take advantage of the commit to change the name of the `Build status` badge in
the `README` file.