ENH: Update `ISMRM2015` submission data
Update `ISMRM2015` submission data.

The previous data, obtained by copying the data available under the form
of dynamic tables in the former version of the `tractometer` website,
contained errors due to oversights from the authors. Thus, none of the
scores corresponded faithfully to the submissions and the published
Nature Communications paper.

Part of the original data, corresponding to the bundle-wise and global
bundle-related scores, faithful to the submissions, was retrieved and
made available by members of the SCIL lab in September, 2022. Yet, the
angular data was not retrieved/made available with that release, so the
old data is kept around for the individual submission files, and an "NA"
value is added for the global data until the data is retrieved.

See: https://tractometer.org/ismrm2015/results/