ENH: Refactor the dataset fetchers
Refactor the dataset fetcher to for an improved code re-use:
- The dataset tuples no longer provide an effective fetcher, but only
the parameters needed to make the fetcher. The `fetch_` prefix is
removed from their names as well. Add a string to the tuples
corresponding to the `msg` parameter of the `_make_fetcher` function and
swap the items corresponding to the `doc` and `data_size` parameters so
that when the `_make_fetcher` function is called all parameters are
provided in the appropriate order.
- Use a method to provide the fetcher parameters depending on the
dataset name rather than having the control statements in the filename
provider method.
- The fetcher names are all built in the same way, and thus this is put
into a method for the sake of best coding practices. Also, prefer naming
the fetchers directly using the Dataset enum values.

Simplify the filename provider method accordingly.

Remove "Download" from the dataset description parameters as they
no longer provide a fetcher in itself, and they rather describe the
dataset itself.

Test the newly added methods.